refactor: reorganize admin endpoint classes by domain
Rename AdminTeamsEndpoints -> AdminSlackEndpoints and AdminIndexingEndpoints
-> AdminSearchEndpoints to match the admin UI's nav grouping. Move the
Slack/Discord broadcast handlers out of the standalone AdminBroadcastEndpoints
into AdminSlackEndpoints/AdminDiscordEndpoints respectively so each admin
endpoint class owns one integration end-to-end. No route changes.
