--- src/FplBot/Services/WebApi/Endpoints/Api/Admin/AdminBroadcastEndpoints.cs
@@ -1,47 +1,0 @@
-using FplBot.Messaging.Contracts.Commands.v1;
-using MassTransit;
-
-namespace FplBot.WebApi.Endpoints.Api.Admin;
-
-public record BroadcastRequest(string Message);
-
-public record DiscordBroadcastRequest(string Message, ChannelFilter Filter);
-
-public static class AdminBroadcastEndpoints
-{
-    public static void Map(RouteGroupBuilder group)
-    {
-        group.MapPost("/slack/broadcast", BroadcastToSlack);
-        group.MapPost("/discord/broadcast", BroadcastToDiscord);
-    }
-
-    private static async Task<IResult> BroadcastToSlack(BroadcastRequest request, ISendEndpointProvider sendEndpointProvider, ILogger<Program> logger)
-    {
-        logger.LogInformation("ENQUEUEING BROADCAST TO SLACK");
-        try
-        {
-            var endpoint = await sendEndpointProvider.GetSendEndpoint(new Uri($"queue:{nameof(EventHandlers.Slack.BroadcastToSlackHandler)}"));
-            await endpoint.Send(new FplBot.Messaging.Contracts.Commands.v1.BroadcastToSlack(request.Message));
-            return TypedResults.Ok(new { message = "Slack Broadcast enqueued!" });
-        }
-        catch (Exception e)
-        {
-            return TypedResults.Ok(new { message = $"Broadcast to Slack failed '{e}'" });
-        }
-    }
-
-    private static async Task<IResult> BroadcastToDiscord(DiscordBroadcastRequest request, ISendEndpointProvider sendEndpointProvider, ILogger<Program> logger)
-    {
-        logger.LogInformation("ENQUEUEING BROADCAST TO DISCORD");
-        try
-        {
-            var endpoint = await sendEndpointProvider.GetSendEndpoint(new Uri($"queue:{nameof(EventHandlers.Discord.BroadcastHandler)}"));
-            await endpoint.Send(new FplBot.Messaging.Contracts.Commands.v1.BroadcastToDiscord(request.Message, request.Filter));
-            return TypedResults.Ok(new { message = $"Discord Broadcast enqueued using {request.Filter}!" });
-        }
-        catch (Exception e)
-        {
-            return TypedResults.Ok(new { message = $"Broadcast to Discord failed '{e}'" });
-        }
-    }
-}
